payment_project/requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
ccxt==4.5.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.6
Django==5.2.5
django-ninja==1.4.3
frozenlist==1.7.0
idna==3.10
kombu==5.5.4
loguru==0.7.3
multidict==6.6.4
packaging==25.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
pycares==4.10.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.4.0
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
win32_setctime==1.2.0
yarl==1.20.1
